Backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# Database (SQLite - zero cost)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

# ===========================================
# AWS SDK (PRIMARY - Hackathon)
# ===========================================
boto3==1.34.49
botocore==1.34.49

# ===========================================
# FALLBACK: AI/ML Services (All Free)
# ===========================================

# LLM Fallbacks
google-generativeai==0.4.0
openai==1.12.0

# Audio Processing (Free/Local)
openai-whisper==20231117
pydub==0.25.1
SpeechRecognition==3.10.1

# Translation (Free)
deep-translator==1.11.4
langdetect==1.0.9

# Vision (Free/Local)
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4

# ===========================================
# Storage & Social (NEW)
# ===========================================
firebase-admin==6.4.0
APScheduler==3.10.4
twikit==1.5.0                  # Twitter/X (No API Key)
linkedin-api==2.1.0            # LinkedIn Service

# ===========================================
# Utilities
# ===========================================
cachetools==5.3.2
python-dateutil==2.8.2
tenacity==8.2.3

# ===========================================
# Testing
# ===========================================
pytest==8.0.1
pytest-asyncio==0.23.5

# Content Moderation (Local/Fallback)
localmod @ git+https://github.com/KOKOSde/localmod.git

